server/requirements.txt
cryptography==36.0.1
Flask==2.0.2
psycopg2==2.9.3
PyJWT==2.4.0
pytz==2023.3
requests==2.31.0